Separate redis hash tests
